Define `hindent`
